+++ js/src/tests/missingThings.test.js
@@ -1,11 +1,11 @@
-let allTheThings = require( "../researches/allTheThings" );
+let allTheThings = require( "../researches/missingThings" );
 let Paper = require( "yoastseo/js/values/Paper.js" );
 
 describe( "This will check if you've written all the things people need to know!", function(){
 	it ( "returns the moonlanding as the next thing to write about", function(){
 		let mockPaper = new Paper( "This is an ordinary text." );
-		let thing = allTheThings( mockPaper );
-		expect( thing ).toEqual( "moon landing" );
+		let nextThing = allTheThings( mockPaper );
+		expect( nextThing ).toEqual( "moon landing" );
 	});
 
 	it ( "returns flat earth as the next thing to write about", function(){
@@ -14,7 +14,7 @@ describe( "This will check if you've written all the things people need to know!
 		expect( thing ).toEqual( "earth is flat" );
 	});
 
-	it ( "returns ufo as the next thing to write about", function(){
+	it ( "returns ufo as the next thing to write about, because this is the first thing in the list that is missing.", function(){
 		let mockPaper = new Paper( "The moon landing was staged, the earth is flat and run by the reptilians " );
 		let thing = allTheThings( mockPaper );
 		expect( thing ).toEqual( "ufo" );
